"aprgcommon refactoring"

+++ AprgCommon/AprgCommon/src/File/AlbaFileReader.cpp
@@ -3,6 +3,7 @@
 #include <Bit/AlbaBitManipulation.hpp>
 #include <String/AlbaStringHelper.hpp>
 
+#include <cstdint>
 #include <fstream>
 #include <numeric>
 #include <string>
@@ -94,7 +95,9 @@ NumberType AlbaFileReader::getData()
     NumberType result(0);
     m_stream.read(getCharacterBufferPointer(), numberOfBytesToRead);
     unsigned int numberOfCharacters = static_cast<unsigned int>(m_stream.gcount());
-    result = accumulate(m_characterBuffer.cbegin(), m_characterBuffer.cbegin()+static_cast<int>(numberOfCharacters), static_cast<NumberType>(0U), [&](NumberType partialSum, NumberType newValue)
+    result = accumulate(m_characterBuffer.cbegin(), m_characterBuffer.cbegin()+static_cast<int>(numberOfCharacters),
+                        static_cast<NumberType>(0U),
+                        [&](NumberType partialSum, NumberType newValue)
     {
         partialSum = static_cast<NumberType>(partialSum << 8);
         partialSum |= (0xFF & newValue);
@@ -103,15 +106,21 @@ NumberType AlbaFileReader::getData()
     return result;
 }
 
-template unsigned int AlbaFileReader::getTwoByteData<unsigned int>();
-template unsigned int AlbaFileReader::getFourByteData<unsigned int>();
-template unsigned long long AlbaFileReader::getEightByteData<unsigned long long>();
-template unsigned int AlbaFileReader::getTwoByteSwappedData<unsigned int>();
-template unsigned int AlbaFileReader::getFourByteSwappedData<unsigned int>();
-template unsigned long long AlbaFileReader::getEightByteSwappedData<unsigned long long>();
-template unsigned int AlbaFileReader::getData<unsigned int, 2>();
-template unsigned int AlbaFileReader::getData<unsigned int, 4>();
-template unsigned int AlbaFileReader::getData<unsigned int, 8>();
+template uint16_t AlbaFileReader::getTwoByteData<uint16_t>();
+
+template uint32_t AlbaFileReader::getFourByteData<uint32_t>();
+
+template uint64_t AlbaFileReader::getEightByteData<uint64_t>();
+
+template uint16_t AlbaFileReader::getTwoByteSwappedData<uint16_t>();
+
+template uint32_t AlbaFileReader::getFourByteSwappedData<uint32_t>();
+
+template uint64_t AlbaFileReader::getEightByteSwappedData<uint64_t>();
+
+template uint16_t AlbaFileReader::getData<uint16_t, 2>();
+template uint32_t AlbaFileReader::getData<uint32_t, 4>();
+template uint64_t AlbaFileReader::getData<uint64_t, 8>();
 
 void AlbaFileReader::saveDataToMemoryBuffer(AlbaMemoryBuffer& buffer, unsigned int numberOfBytesToRead)
 {
